ENH:Fixed float to double conversion leftover. COMP:Fixed warnings.

+++ Rendering/vtkVolume.h
@@ -219,7 +219,7 @@ protected:
   vtkTimeStamp         GradientOpacityArrayMTime[VTK_MAX_VRCOMP];
 
   // Function to compute screen coverage of this volume
-  float ComputeScreenCoverage( vtkViewport *vp );
+  double ComputeScreenCoverage( vtkViewport *vp );
   
 private:
   vtkVolume(const vtkVolume&);  // Not implemented.
